GH actions: Pin 20.04 image for now, to support systemd-in-a-container
The /sys/fs/cgroup mount looks noticably different in 22.04 on GH actions,
as shown by a "ls -lisa /sys/fs/cgroup" (amongst other details, it does
no longer include a "systemd" entry), which is not compatible with an
older Ubuntu systemd running in a container.

Its not clear whether a newer Ubuntu in the container would fix this, or
more investigation is needed, but for now pin Ubuntu on the host to 20.04
to ensure /sys/fs/cgroup is compatible.

For future reference, a failure here shows as systemd silently not
starting, without giving any error messages, leading to a confusing Docker
error message since the container immediately returns.

In passing, reduce the mountpoints we take from the host, as only
/sys/fs/cgroup is actually needed. Mark this remaining mount as readonly,
for good measure.