GH actions: Update integration tests for Ubuntu 22.04 and pin it
There is a distinct difference between cgroup v1 and v2, which the
GH action change of "ubuntu-latest" (step-by-step migration to 22.04)
caused us to run into.

For consistency, and because it does actually matter which version we
are on, pin 22.04 and adjust the tests accordingly. With cgroups v2
we can now rely on giving the integration test containers their own
cgroup namespace, instead of bind mounting the host's /sys/fs/cgroup.

To support this change on GH actions we also adjust the reload task
to use journalctl for checking collector log messages, instead of
utilizing a file in the temp folder, that was causing odd failures that
could only be reproduced in the GH actions environment.