Improve Aiven Service ID and Project ID detection from hostname
Currently, we assume that Aiven hostnames follow the pattern

    {service-id-part1}-{service-id-part2}-{project-id}.aivencloud.com

This is the case for default service names, but a service name does
not need to be dash-separated.

Update our detection logic to account for arbitrary project names that
follow Aiven's documented pattern

    <SERVICE_NAME>-<PROJECT_NAME>.aivencloud.com

Assume that project names do not contain dashes (setting
aiven_project_id and aiven_service_id manually is still an option if
that's not the case).