Support IAM token authentication for Amazon RDS and Amazon Aurora
This adds a new configuration setting, db_use_iam_auth / DB_USE_IAM_AUTH
that, if enabled, causes the collector to fetch a short-lived token for
logging into the database instance from the AWS API, instead of using a
hardcoded password in the collector configuration file.

In order to use this setting IAM authentication needs to be enabled on the
database instance / cluster, and the pganalyze IAM policy needs to be
extended to cover the "rds-db:connect" privilege for the pganalyze user:

https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/UsingWithRDS.IAMDBAuth.IAMPolicy.html

Note that for now the collector will fetch a new token for each connection
it makes (at least one new connection every 10 seconds), which may cause
additional AWS API utilization when monitoring a lot of instances.

With thanks to Alexandre Viau for the initial code contribution.