AWS: Avoid DescribeDBInstances calls for log downloads for RDS instances (#369)
Previously the log download for RDS would issue a DescribeDBInstances call
for confirming whether the instance ID is indeed valid.

Whilst this was made less problematic by addition of a cache in
f2ab72f45bd, it missed the fact that we don't actually need to do this
lookup at all in the case of log downloads for Amazon RDS instances.

Instead, we can just rely on the later DescribeDBLogFiles API call to
complain if an instance ID does not exist. Thus, this change removes that
API call completely for log downloads when the instance ID is specified.

In the case of Aurora we do need the API call to DescribeDBClusters to
find the specific instance ID, e.g. that of the current writer. In terms
of caching we can't cache the cluster ID => instance ID mapping for risk
of missing a failover event (i.e. where the cluster ID should start
pointing to a new instance ID right away), but we can cache errors, e.g.
when someone specified an incorrect cluster ID.

With thanks to Jesse Liston for pointing out the unnecessary
DescribeDBInstances call.