Amazon RDS/Aurora: Improve handling of excessively large log file portions
We've previously introduced a limit of 10 MB of log data per file tracked
in a log snapshot, to avoid causing log analysis to time out in cases of
very large log events, or other issues that caused a lot of data to
be written since the last log marker.

However, the prior logic was still keeping the full original log data
around *and* unnecessarily encrypted and submitted that data, even though
it was not actually usable. Additionally, the prior logic was hard to
reason about, and had subtle bugs, such as the one fixed in 90db3291ec.

Instead, in cases of the downloaded log data exceeding 10 MB, write out
a new temporary file that contains just that data. In addition, add a
warning in the collector logs to highlight when this happens, so that
a user may take action (e.g. reduce occurrence of problematic log events).

In passing this refactors the RDS log handling code into more targeted
functions for clarity, as well as potentially helping Go enforce scoping
rules better for the individual downloaded log file portion (to ensure
the downloadd log text is not kept in memory after its been written to
the temp file).