Log parsing: Only consider first 1000 characters for log_line_prefix
Sometimes log lines can be excessively long, but the log_line_prefix is
almost always a tiny portion of such log lines. Since we are currently
doing Regexp-based parsing for log lines (before analysis), it can be
expensive to pass such large log lines in full to the Go regexp engine.

Instead, only pass the first 1000 characters to the Regexp engine when
reading out the log_line_prefix fields. In a real world test case this
improves the parsing performance for a 2.1GB log file from 17 minutes
to 14 minutes (tested with "time pganalyze-collector --analyze-logfile"),
and its hard to imagine a case where the prefix is longer than 1000
characters.