Relation stats: Correctly filter out special tables for constraints query
Previously the constraints query included foreign tables, unlike the
earlier query for gathering relation info that excluded them. This then
causes a bogus empty table to show up in the snapshot, due to Go
initializing an empty struct when looking up an OID that does not exist.

Fix by aligning the constraints query with the other queries, and
ignoring the same kind of tables.