Update from github.com/golang/protobuf to google.golang.org/protobuf
The old module, github.com/golang/protobuf, has been deprecated a while,
and the newer API has a slightly improved interface for the timestamp
type as well, avoiding a potential error that needs to be handled (that
we've traditionally ignored anyway in our own code).

Note there is a remaining caller of the old module in Google's appengine
module, which gets vendored in unnecessarily via x/oauth2 (https://github.com/golang/oauth2/issues/615).

In passing, update the protoc-gen-go tool as well to the version hosted
in the new repository.

Depends on https://github.com/pganalyze/collector-snapshot/pull/30