Pass down context to via runners to input/output functions and use it
This ensures that cancellations can be propagated correctly, and prepares
for future commits that introduce additional cases where context is used
for cancellation.

Note this is overall mostly a mechanical change, with two exceptions:

1) The collection of schema statistics for all databases can now fail
   and return an error (in case of a context being cancelled) - previously
   errors here were treated as database specific and only logged, and we
   continue to do so for anything that's not an error in the caller's
   context

2) Improved error handling in CollectFull, RunExplain, and connection
   establishment functions to ensure context errors are reported

Additionally, note that some functions do not report errors right now,
e.g. StatsHelperExists, and to avoid further code churn these are left
as is for now. The assumption is that a context error would be picked
up by the next caller for now.