Allow SIGINT to interrupt collector test commands
This uses the recent context propagation improvements to stop the
collector right away in case Ctrl+C is pressed during a test run
(i.e. SIGINT is sent to the process).

Note that this also allows the long-running process (i.e. not in a test
run) to be stopped with SIGINT going forward, not only SIGTERM - but
that seems acceptable.