Schema data collection: Use at most 8 minutes to collect schema data
Previously we enforced a per-database timeout when collecting schema
data, as a side effect of having a statement timeout, but there was no
server-wide timeout.

In some reported cases this caused the collector to spend a long time
with particular servers (e.g. if there are thousands of databases),
avoiding the scheduled runs for other servers, and causing bad data to
be reported.

To avoid this, add an explicit timeout of 8 minutes, which gives the
other collector queries 2 minutes to do their work in case schema
statistics are taking that full amount.

Note the chosen timeout here is essentially "a reasonable number
that's less than 10 minutes", with the idea that other functions run by
the collector (e.g. retrieval of statement statistics) can often complete
in under 2 minutes total, even on a server with many databases, since
they are only run once for the whole server.