Scheduler: Enforce time limit for all scheduled runs based on interval
This adds a context-based deadline that avoids an individual run
exceeding the time allocated (i.e. the interval for it's group).

For example in the case of full snapshots, this will enforce them to run
at most for 10 minutes, avoiding edge cases where they can take longer.

Note that this also applies to high frequency snapshots, such as activity
snapshots (that are now only allowed < 10 seconds to complete).

For now this is not applied to secondary groups in ScheduleSecondary
(typically the runner for gathering query stats without text) because
it would cause an unexpected failure with "context deadline exceeded"
when the primary group takes longer than the interval of the seconary
group (i.e. full snapshot taking longer than 1 minute).

In passing this removes unnecessary "return" statements, and removes
a deprecated notice about how the scheduler might work in the future
(the fixed interval scheduling is pretty established now, and not going
away anytime soon).