Fix Azure Single and Flexible Server log line parsing for DETAIL lines
The code previously did not work correctly since the log_line_prefix
was not present on the DETAIL lines (which are included on the main
log line on Azure). This subsequently caused issues with log-based EXPLAIN
not having DETAIL lines available.

Fix by reworking how log lines get transformed for Azure, and adding
unit tests for the transformation based on examples from running
"pgbench -M extended".