Relation queries: Correctly handle later queries encountering a lock
The existing coding assumed that we can rely on the fact that a relation
is exclusively locked to stay the same across each query that works with
relation data to get columns, indexes, constraints, etc.

But that is not true (locks can change between the queries) and that
causes a race condition where the first query does not see a relation
being locked, but later queries do. This then causes downstream problems
since only part of the information was sent, and e.g. an index might
show in one snapshot, and then disappear in the next, if locking
conditions are unlucky.

To fix, return which exclusive locks each relation query encountered, and
take that to consider the relation having been exclusively locked for the
whole current snapshot, and set the flag accordingly.

In passing, add explicit checks on subsequent relation queries to
ensure the relation actually was captured in the first query. This is
to protect against locks being held on relations that we ignored earlier
(e.g. because of them being a catalog table, or matching the schema filter)
suddenly showing up with empty data through a later query returning them
as a locked relation.

Note that the relation statistics logic suffers from a different but
related problem which is to be fixed in a subsequent commit.