Relation statistics: Avoid bogus data due to diffs against locked objects
Previously we would skip locked relations when getting the statistics for
either the table or an index. Compared to the locking race condition
in the relation info queries, this is actually not directly a problem in
the same way.

However, the problem with the stats queries is that they assume they
can diff against the prior record, and if that doesn't exist, that the
table (or index) was recently added, and thus we can diff against zero.

That is not correct when relations were ignored due to locks, leading
to a subsequent stats collection suddenly seeing the full counter value
as the diff (since the prior data point was skipped due to the lock).

Fix by explicitly tracking locked relations that the stats queries
skipped, and applying that knowledge in the diff handling.