Amazon RDS / Aurora: Add new "rds-ca-global" option for db_sslrootcert
We've historically bundled the individual CAs certs that AWS publishes,
to help use a verify-full connection with RDS. With the recently released
set of new CA options, i.e. "rds-ca-rsa2048-g1", "rds-ca-rsa4096-g" and
"rds-ca-ecc384-g1", AWS has instead started publishing a combined list
of CAs in a single "global" file. This global set also includes the
old "rds-ca-2019-root" (but not the 2015 root, since that is expired).

To simplify configuration, introduce the new "rds-ca-global" option for
db_sslrootcert, which is the recommended configuration for RDS and Aurora
going forward. For compatibility reasons we still support naming the
"rds-ca-2019-root" explicitly, but its now just an alias for the global
set. This also drops the "rds-ca-2015-root", which would no longer have
worked anyway (since the cert expired).