Query stats reset: Make pg_stat_statements_reset() failure a soft error
Previously this was a hard error, which failed the snapshot - this is
not desirable because hard errors mean the snapshot state does not get
persisted. In practice this meant that the error would very likely occur
again on the next full snapshot run, in the typical case of a permission
problem. This meant that full snapshots kept failing, which indirectly
led to a memory leak because the UnidentifiedStatementStats list would
keep growing (as the query runner was working just fine, once a minute).

To fix, emit only a warning instead, and process the rest of the snapshot
as usual, thus persisting the state and not calling the reset function
again until the next scheduled interval.

Fixes #398