Scheduler: Use 19 seconds to check for very short deadlines
This makes it easier for a future reader to understand, since it matches
the if condition with the override we're setting.

In practice this doesn't actually matter with the current scheduler config
since we either have 10 seconds or 1 minute apart, and thus the 10
seconds interval would already have matched as a very short deadline.