Build collector packages using GitHub Actions (#397)
This simplifies the release process by utilizing GH actions to build
the RPM/Apt packages. Note that this only includes the build and test
steps - the signing and release continue to be done separately.

Additionally we now build and test packages once each day. This aims
to reduce the chance of a test failure late in the release process.