RPM packages: Ensure collector gets started after reboot
This was an oversight in the packaging logic, since we utilize fpm's
"--deb-systemd-auto-start" setting for the Debian-based packages, but
we don't have the equivalent for the RPM packages.

Fix by adding an explicit "systemctl enable" on installation, and a
corresponding "systemctl disable" on removal, matching what was already
done for the Debian packages.