Heroku: Replace use of lpx module with direct code to parse log lines
Whilst the code in "github.com/bmizerany/lpx" is trivial, and from our
relationship with the author we know that its not a problem to use it,
its technically not licensed under a publicly documented OSS license.

To avoid any potential confusion by license check tools, replace "lpx"
with similar logic inline in the Heroku log parsing logic.