Log analysis: Remember unqualified name for "skipping vacuum" events
Due to the way we associate log lines to relation references, we need to
have a fully qualified name that has both a schema and relation name.

Unfortunately Postgres currently never outputs a schema name for the
"skipping vacuum" and "skipping analyze" events, only the table name
itself - which caused us to not associate the line to a relation.

To allow filtering on the application side through special logic that
can look at tables in all schemas matching the name, remember the
relation name in the unstructured details field of the log line, and stop
storing it in the RelationName field (which was never sent to the server
anyways).