Use log_timezone setting to determine log timezone if possible
Right now, we infer the log timezone by parsing this out of log lines.
This works in theory, but in practice Go's time.LoadLocation does not
support many timezone abbreviations (and abbreviations can be
ambiguous anyway).

Since we already collect Postgres settings, and these specify the log
timezone in a non-abbreviated form, we can use these. These correspond
to the full name in the IANA Time Zone database, as time.LoadLocation
expects. Sync this on every full snapshot just in case the timezone
is updated.