AWS Aurora: Correctly detect Aurora reader instances as replicas
Previously we skipped processing of replication statistics for Aurora,
since the regular built-in Postgres replication functions error out in
many cases. However, this also meant that all Aurora instances were
always detected as primaries, regardless of whether they were a writer or
a reader, which led to confusing in-app experiences.

Fix by adding Aurora-specific logic in the replication stats collection.

This also adds support for ignoring Aurora readers using the
"skip_if_replica" collector configuration option. This can be used when
specifying Aurora instances individually as servers, but only intending
to monitor the writer. Note that for such situations we do generally
recommend specifying the cluster write endpoint instead, but there is no
downside to support this alternate way of only monitoring writers.