Docker log tail: Re-order args to also support podman aliased as docker
Whilst we currently explicitly call out to "docker" for this experimental
way of reading local logs during development, we technically also support
podman when its available with the "docker" alias. However, podman
requires arguments to be set before the container name. Fix support for
this type of setup. This keeps compatibility with docker, which accepts
arguments in either order.