Integration tests: Ensure en_US.UTF-8 locale remains installed after locale-gen
See https://github.com/docker-library/postgres/issues/1112, this is necessary
because the Docker Postgres images directly call "localedef", which then gets
overriden by a later locale-gen. locale-gen reads from /etc/locale.gen as the
authorative source, and removes any manually defined locales.

This situation can easily occur when the "locales" package is upgraded due to
an apt-get install that depends on the package, as is done in the integration
tests.