Activity snapshots: Normalize queries for "filter_query_sample = normalize"
For historic reasons, we were only normalizing pg_stat_activity query
texts when "filter_query_sample" was set to "all", not "normalize", which
has a similiar intent.

Whilst we could also control the filtering of pg_stat_activity query texts
with the "filter_query_text" setting (which today controls whether
pg_stat_statements query texts are re-normalized), this is intentionally
not done here. The main motivation to not do this for now are performance
implications of running normalize every 10 seconds for many active queries,
and we may adjust this in a future commit based on more benchmarks.