Improve handling of incorrect shared_preload_libraries setting
We used to special case this to avoid the collector not reporting data,
but since we just changed how we handle pg_stat_statements errors (we
no longer fail the snapshot), we can simplify the logic a bit.

However, since its actually useful to know when this situation occurs,
add another HINT message that explicitly points to a missing server
restart as being a possible issue.