Fix Citus integration test (#489)
The current test was running without the Citus extension, since we
were passing stdin to `docker exec`, to pass to psql, but without
running it in interactive mode. Docker discarded stdin, and the
test was running without Citus-specific functionality.

Pass the -i flag to `docker exec` so it relays stdin to psql.  Also
create the Citus extension (which is not created by default in the
Citus Docker image), and update the expected output file.