Avoid error collecting schema stats on Citus tables with no indexes (#488)
The Citus `run_command_on_shards` function we use to collect
distributed index stats appears to return an empty string when the
query passed in returns zero rows. We currently try to cast that to
jsonb, which fails (this is not a valid jsonb value).

Ignore any empty strings returned by `run_command_on_shards`.