Warn in guided setup integration test when collector warnings present (#497)
We don't expect any warnings and should confirm that's the
case. Warnings will often also cause a test failure, but that can be
hard to track down since the collector log is a number of lines
earlier than the diff failure output, and there's a lot of output in
integration tests.