Self-hosted: Allow collector to receive logs via OpenTelemetry
This is mainly intended to support Kubernetes, by using fluentbit with the
opentelemetry output plugin to send pod logs to the collector. The logs
received can optionally be filtered by pod name, or via label selectors.

Note this requires the log output to be jsonlog (Postgres 15+), with
optional K8s context added, or plain logs without any additional context.

An example fluentbit output configuration:

  [OUTPUT]
    name  opentelemetry
    match kube.*postgres*
    host  127.0.0.1
    port  4318

With a corresponding collector configuration:

  db_log_otel_server = 127.0.0.1:4318

When utilizing CloudNativePG, a filter like the following can be used
to filter only to the primary's logs for a given cluster:

  db_log_otel_k8s_labels = cnpg.io/cluster=cluster-example,cnpg.io/instanceRole=primary