Relation stats: Call pg_stat_get_* directly instead of using system views
Due to the way how pg_stat_all_tables and pg_statio_all_tables are
defined, its significantly more expensive to query through the views
when a table filter is in place, since all the data gets pulled, only
for the filter to be applied afterwards.

Instead, call the underlying pg_stat_get* functions directly, which has
the same effect as querying the views (as they are simple views without
any security barrier), and shows a major speedup in cases where a large
portion of tables is filtered out. This also should provide a minor
speedup in regular operations, even with no table filter in place.