Amazon RDS/Aurora IAM Authentication: Default to DB port 5432
Previously IAM authentication would fail with "PAM authentication failed"
when the port was not explicitly set in the collector configuration.

This instead falls back to the default (5432), which is the typical port
used on Amazon RDS and Aurora.

In passing use the same logic in the fallback logic for identifying
RDS instances by their hostname, since the DB instance list also always
includes the port, and this was likely broken as well if the port wasn't
set explicitly.