Add --generate-helper-sql command to generate helpers psql install script
Passing the new "--generate-helper-sql" command to the collector will
generate a SQL script that can be passed to "psql" to install helpers
(e.g. for collecting column stats) on all configured databases for the
specified server.

The command takes the name of the configuration section of a specific
server in the collector configuration (e.g. "server1").

It can be used manually by saving the output to a file, or by piping
it to psql like this:

pganalyze-collector --generate-helper-sql=server1 | psql -f -

Make sure to execute the generated SQL as an administrative user or
database owner (not the pganalyze user!) in order for the
SECURITY DEFINER functions to have sufficient privileges assigned.