Add --generate-stats-helper-sql command to generate helpers psql install script (#535)
Passing the new "--generate-stats-helper-sql" command to the collector will
generate a SQL script that can be passed to "psql" to install stats helpers
(e.g. for collecting column stats) on all configured databases for the
specified server.

The command takes the name of the configuration section of a specific
server in the collector configuration (e.g. "server1"), or "default" when
using environment variable-based configuration.

It can be used manually by saving the output to a file, or by piping
it to psql like this:

pganalyze-collector --generate-stats-helper-sql=server1 | psql -f -

Make sure to execute the generated SQL as an administrative user or
database owner (not the pganalyze user!) in order for the
SECURITY DEFINER helper functions to have sufficient privileges assigned.