Restore "Setting up a Restricted Monitoring User" section (#533)
Use it to point to our official docs. Users may still be pointed here
from, e.g., log output of old collector versions, so we can make it
easier to find the right documentation.

Also drop the Additional Setup parts of README

This is for functionality that is no longer supported.