Work around RDS pg_stats and pg_stats_ext issue
It looks like RDS has some custom patches that can run the permission
check for stats before running the related filter, causing a
permission failure fetching stats for tables that we don't want to see
in the first place (and that should be pruned by the `schemaname`
filter).

Ignore the 'pg_subscription' table explicitly.