Test run: Avoid collector hanging when interrupted via CTRL+C (SIGINT)
Due to the way the signal handling is written, we stop checking the
testRunSuccess channel once a SIGINT is received. However the code
that marks the wait groups as done could only return once it finished
writing to the channel, since it was unbuffered. Make the channel buffered
so that the code can return as intended.

Fixes #561