Fix DISABLE_CITUS_SCHEMA_STATS config reading (#568)
The changes in #562 missed that the updated config-reading code was
only for config read from environment variables. Config from the
config file was not pre-processed, so any currently-documented values
of disable_citus_schema_stats specified there would be ignored, and
relation and index stats would always be collected. Config specified
via individual environment variables worked as expected.

Since the new buggy behavior of the setting is not yet in a released
version, and not yet documented, it does not need a CHANGELOG entry.