Add alternate mechanism for parsing log lines (#494)
Instead of supporting a list of predefined `log_line_prefix`es, and testing
each log line against each prefix, check `log_line_prefix` at startup (and
watch for changes during full snapshots), and build a custom regex based
on the prefix. This is faster and more flexible, since it can support arbitrary
prefixes.

Add a `db_log_line_prefix` option for now to fall back to the old behavior
in case there are issues with the new mechanism. The old parsing code
and this option can be dropped in a future patch.