Collect system info and metrics for Azure (#591)
Collect system info and metrics for Azure Flexible Server and Cosmos DB for PostgreSQL.
To start using this, the customer needs to supply a new config variable AZURE_SUBSCRIPTION_ID to the collector, as well as setting up the managed identity. When AZURE_SUBSCRIPTION_ID is not supplied, the collector doesn't error out and simply won't collect the system info.