Improve support for EDB Postgres Advanced Server (EPAS)
Previously the collector would emit system catalogs that are only existing
on EPAS, causing unnecessary information to show in the pganalyze dashboard.

Additionally, EPAS allows function signatures to be re-used between functions
and procedures (presumably this was allowed before upstream Postgres added
CREATE PROCEDURE), which causes a server-side error with pganalyze. Since this
is an edge case and there are no intents to support this upstream, ignore
duplicate functions/procedures that have the same signature.