Fix --analyze-logfile flag
In #494, I missed updating the --analyze-logfile flag. The flag is
intended to analyze a logfile "offline", which does not work with the
new mechanism, since that relies on being able to read the server's
log_timezone and log_line_prefix.

This adds additional flags to pass that information to the
--analyze-logfile path. The timezone defaults to UTC; the prefix is
required.