Handle the database / user being empty
It is possible for the username and the database to be empty for some log messages. For example with CloudSQL the vacuum analysis logs are output as

> 2024-10-04 06:38:28.316 UTC [2569017]: [1-1] db=,user= LOG:  automatic vacuum of table \\\"database_name.pg_catalog.pg_database\\\": etc

This previously caused a bug where these lines were being treated as incomplete, resuling in a memory leak as "the rest" of the line was awaited.

This commit fixes this by changing the regex matching for username and database to match anything that is not a single space, rather than just anything.