Handle ambiguous log lines more reliably (#606)
In some cases, it's possible for the user, database, or other free-form parts of
the log line prefix to match too much of the input string, leaving the LogLevel
to unintentionally match part of the content.

E.g., with CloudSQL the vacuum analysis logs are output as

> 2024-10-04 06:38:28.808 UTC [2569017]: [1-1] db=,user= LOG: automatic vacuum of table \"some_database.some_schema.some_table\": index scans: ...

This is matched as user " LOG: automatic vacuum of table \"some_database.some_schema.some_table\": index"
and log level "scans". The bogus log level is then mapped to 0, and the line is not
handled correctly.

To avoid this, change our regexp to only look for the log levels (and additional
context strings that appear in place of log level on detail lines) we recognize.