Track cluster identifier (cluster ID) as part of system information (#616)
* Track cluster identifier (cluster ID) as part of system information

This is intended to help identify servers that belong together, to allow
the pganalyze app to group them in an overview, offer data in combined
views, or allow Index Advisor to avoid considering indexes unused if
they are used on a replica.

To start with this identifies the cluster ID for given providers:
- AWS: DBClusterIdentifier (user-assigned cluster name)
- Azure: Source Server ID (replication parent server ID + resource group)
- Crunchy Bridge: Cluster Parent ID (replication parent server ID)
- Self hosted: pg_controldata system identifier (unique ID shared by all physical replicas)

There are providers (e.g. GCP) and edge cases (e.g. when helper doesn't
run for self hosted) where this doesn't work. Handling those cases, and
allowing manual overrides of the cluster ID is left for a later follow-up.