Implement new pganalyze.explain_analyze() helper, refactor query run
This executes on-demand query runs for the Query Tuning feature via
a helper function that prevents reading the table data directly, as well
as granting the collector user unnecessary permissions.

This is now the only way that query runs can execute EXPLAIN ANALYZE,
direct execution without the helper is intentionally not supported.

In passing, expand the protobufs to pass down parameters and parameter
types and optional planner settings, and refactor the query run code to
handle work requiring a DB connection in a separate Go function.

Further, this adds tests that utilize a new TEST_DATABASE_URL enviroment
setting, which is used to run unit tests that require a database
connection (previously only the integration tests used the database).