DB connections: Don't limit to 30 seconds lifetime to avoid timeout bugs
Some time ago (at least 9+ years) we introduced a maximum connection
lifetime of 30 seconds via the SetConnMaxLifetime database/sql function.

This was likely intended as an extra safety measure in case the collector
didn't clean up its DB handle correctly. We've had a few minor issues
with incorrect cleanups over the years (though usually not with the DB
connection), but at this point our connection cleanup should be reliable.

The connection lifetime causes an issue because it causes Go to
reconnect to the database after 30 seconds have elapsed. This then
effectively removes any connection local settings, in particular
a statement_timeout.

Effectively this meant that queries may unintentionally not have been
subject to the statement_timeout, if the same connection previously took
30 seconds to do its work. To resolve, remove the max lifetime.