PII Filtering: Detect bind parameters in CONTEXT as statement_text
This can occur in certain cases, for example auto_explain output
on some Postgres versions. We previously would have not detected
this CONTEXT line, but are now correctly detecting it as containing
statement_parameter log secrets.

In passing add a test case for the new "Query Parameters" field in
auto_explain which gets redacted based on the statement_text filter that
filters out the whole plan text (we normalize the individual fields in
resulting query samples, but the log text for auto_explain is redacted
altogether).